refactor: Break interactive repo component into small components
